editor/libeditor/tests/test_dom_input_event_on_texteditor.html

<html>
<head>
  <title>Test for input event of text editor</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css"
          href="/tests/SimpleTest/test.css" />
</head>
<body>
<div id="display">
  <input type="text" id="input">
  <textarea id="textarea"></textarea>
</div>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(runTests, window);

const kIsMac = navigator.platform.indexOf("Mac") == 0;

function runTests() {
  function doTests(aElement, aDescription, aIsTextarea) {
    aDescription += ": ";
    aElement.focus();
    aElement.value = "";

    var inputEvent = null;

    var handler = function(aEvent) {
      is(aEvent.target, aElement,
         "input event is fired on unexpected element: " + aEvent.target.tagName);
      ok(aEvent instanceof InputEvent,
         "input event should be dispatched with InputEvent interface");
      ok(!aEvent.cancelable, "input event must not be cancelable");
      ok(aEvent.bubbles, "input event must be bubbles");
      let duration = Math.abs(window.performance.now() - aEvent.timeStamp);
      ok(duration < 30 * 1000,
         "perhaps, timestamp wasn't set correctly :" + aEvent.timeStamp +
         " (expected it to be within 30s of the current time but it " +
         "differed by " + duration + "ms)");
      inputEvent = aEvent;
    };

    aElement.addEventListener("input", handler, true);

    inputEvent = null;
    sendString("a");
    is(aElement.value, "a", aDescription + "'a' key didn't change the value");
    ok(inputEvent, aDescription + "input event wasn't fired by 'a' key");
    ok(inputEvent.isTrusted, aDescription + "input event by 'a' key wasn't trusted event");
    is(inputEvent.inputType, "insertText",
       aDescription + 'inputType should be "insertText" when typing "a"');
    is(inputEvent.data, "a",
       aDescription + 'data should be "a" when typing "a"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when typing "a"');

    inputEvent = null;
    synthesizeKey("KEY_Backspace");
    is(aElement.value, "", aDescription + "BackSpace key didn't remove the value");
    ok(inputEvent, aDescription + "input event wasn't fired by BackSpace key");
    ok(inputEvent.isTrusted, aDescription + "input event by BackSpace key wasn't trusted event");
    is(inputEvent.inputType, "deleteContentBackward",
       aDescription + 'inputType should be "deleteContentBackward" when pressing "Backspace" with collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Backspace" with collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Backspace" with collapsed selection');

    if (aIsTextarea) {
      inputEvent = null;
      synthesizeKey("KEY_Enter");
      is(aElement.value, "\n", aDescription + "Enter key didn't change the value");
      ok(inputEvent, aDescription + "input event wasn't fired by Enter key");
      ok(inputEvent.isTrusted, aDescription + "input event by Enter key wasn't trusted event");
      is(inputEvent.inputType, "insertLineBreak",
         aDescription + 'inputType should be "insertLineBreak" when pressing "Enter"');
      is(inputEvent.data, null,
         aDescription + 'data should be null when pressing "Enter"');
      is(inputEvent.dataTransfer, null,
         aDescription + 'dataTransfer should be null when pressing "Enter"');
    }

    inputEvent = null;
    aElement.value = "foo-bar";
    is(aElement.value, "foo-bar", aDescription + "value wasn't set");
    ok(!inputEvent, aDescription + "input event was fired by setting value");

    inputEvent = null;
    aElement.value = "";
    is(aElement.value, "", aDescription + "value wasn't set (empty)");
    ok(!inputEvent, aDescription + "input event was fired by setting empty value");

    inputEvent = null;
    sendString(" ");
    is(aElement.value, " ", aDescription + "Space key didn't change the value");
    ok(inputEvent, aDescription + "input event wasn't fired by Space key");
    ok(inputEvent.isTrusted, aDescription + "input event by Space key wasn't trusted event");
    is(inputEvent.inputType, "insertText",
       aDescription + 'inputType should be "insertText" when typing " "');
    is(inputEvent.data, " ",
       aDescription + 'data should be " " when typing " "');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when typing " "');

    inputEvent = null;
    synthesizeKey("KEY_Delete");
    is(aElement.value, " ", aDescription + "Delete key removed the value");
    ok(!inputEvent, aDescription + "input event was fired by Delete key at the end");

    inputEvent = null;
    synthesizeKey("KEY_ArrowLeft");
    is(aElement.value, " ", aDescription + "Left key removed the value");
    ok(!inputEvent, aDescription + "input event was fired by Left key");

    inputEvent = null;
    synthesizeKey("KEY_Delete");
    is(aElement.value, "", aDescription + "Delete key didn't remove the value");
    ok(inputEvent, aDescription + "input event wasn't fired by Delete key at the start");
    ok(inputEvent.isTrusted, aDescription + "input event by Delete key wasn't trusted event");
    is(inputEvent.inputType, "deleteContentForward",
       aDescription + 'inputType should be "deleteContentForward" when pressing "Delete" with collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Delete" with collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Delete" with collapsed selection');

    inputEvent = null;
    synthesizeKey("z", {accelKey: true});
    is(aElement.value, " ", aDescription + "Accel+Z key didn't undo the value");
    ok(inputEvent, aDescription + "input event wasn't fired by Undo");
    ok(inputEvent.isTrusted, aDescription + "input event by Undo wasn't trusted event");
    is(inputEvent.inputType, "historyUndo",
       aDescription + 'inputType should be "historyUndo" when doing "Undo"');
    is(inputEvent.data, null,
       aDescription + 'data should be null when doing "Undo"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when doing "Undo"');

    inputEvent = null;
    synthesizeKey("Z", {accelKey: true, shiftKey: true});
    is(aElement.value, "", aDescription + "Accel+Y key didn't redo the value");
    ok(inputEvent, aDescription + "input event wasn't fired by Redo");
    ok(inputEvent.isTrusted, aDescription + "input event by Redo wasn't trusted event");
    is(inputEvent.inputType, "historyRedo",
       aDescription + 'inputType should be "historyRedo" when doing "Redo"');
    is(inputEvent.data, null,
       aDescription + 'data should be null when doing "Redo"');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when doing "Redo"');

    // Backspace/Delete with non-collapsed selection.
    aElement.value = "a";
    aElement.select();
    inputEvent = null;
    synthesizeKey("KEY_Backspace");
    ok(inputEvent,
       aDescription + 'input event should be fired by pressing "Backspace" with non-collapsed selection');
    ok(inputEvent.isTrusted,
       aDescription + 'input event should be trusted when pressing "Backspace" with non-collapsed selection');
    is(inputEvent.inputType, "deleteContentBackward",
       aDescription + 'inputType should be "deleteContentBackward" when pressing "Backspace" with non-collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when pressing "Backspace" with non-collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when pressing "Backspace" with non-collapsed selection');

    aElement.value = "a";
    aElement.select();
    inputEvent = null;
    synthesizeKey("KEY_Delete");
    ok(inputEvent,
       aDescription + 'input event should be fired by pressing "Delete" with non-collapsed selection');
    ok(inputEvent.isTrusted,
       aDescription + 'input event should be trusted when pressing "Delete" with non-collapsed selection');
    is(inputEvent.inputType, "deleteContentForward",
       aDescription + 'inputType should be "deleteContentBackward" when Delete "Backspace" with non-collapsed selection');
    is(inputEvent.data, null,
       aDescription + 'data should be null when Delete "Backspace" with non-collapsed selection');
    is(inputEvent.dataTransfer, null,
       aDescription + 'dataTransfer should be null when Delete "Backspace" with non-collapsed selection');

    // Toggling text direction
    aElement.focus();
    inputEvent = null;
    SpecialPowers.doCommand(window, "cmd_switchTextDirection");
    ok(inputEvent,
       aDescription + "input event should be fired by dispatching cmd_switchTextDirection command #1");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when dispatching cmd_switchTextDirection command #1");
    is(inputEvent.inputType, "formatSetBlockTextDirection",
       aDescription + 'inputType should be "formatSetBlockTextDirection" when dispatching cmd_switchTextDirection command #1');
    is(inputEvent.data, "rtl",
       aDescription + 'data should be "rtl" when dispatching cmd_switchTextDirection command #1');
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when dispatching cmd_switchTextDirection command #1");

    inputEvent = null;
    SpecialPowers.doCommand(window, "cmd_switchTextDirection");
    ok(inputEvent,
       aDescription + "input event should be fired by dispatching cmd_switchTextDirection command #2");
    ok(inputEvent.isTrusted,
       aDescription + "input event should be trusted when dispatching cmd_switchTextDirection command #2");
    is(inputEvent.inputType, "formatSetBlockTextDirection",
       aDescription + 'inputType should be "formatSetBlockTextDirection" when dispatching cmd_switchTextDirection command #2');
    is(inputEvent.data, "ltr",
       aDescription + 'data should be "ltr" when dispatching cmd_switchTextDirection command #2');
    is(inputEvent.dataTransfer, null,
       aDescription + "dataTransfer should be null when dispatching cmd_switchTextDirection command #2");

    aElement.removeEventListener("input", handler, true);
  }

  doTests(document.getElementById("input"), "<input type=\"text\">", false);
  doTests(document.getElementById("textarea"), "<textarea>", true);

  SimpleTest.finish();
}

</script>
</body>

</html>
